requirements.txt
# DentAI Pro v2 — Python Dependencies
# ==========================================================
# SETUP:
#   conda create -n dentai python=3.11 -y
#   conda activate dentai
#   pip install -r requirements.txt
# ==========================================================

# ── LangGraph + LangChain ──
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# ── OpenAI ──
openai>=1.30.0

# ── MySQL connector ──
mysql-connector-python>=8.3.0

# ── Environment management ──
python-dotenv>=1.0.0

# ── Streamlit UI ──
streamlit>=1.35.0

# ── Data processing ──
pandas>=2.0.0

# ── Testing ──
pytest>=8.0.0


chromadb>=0.5.0          
sqlalchemy>=2.0.0        
fastapi>=0.111.0        
uvicorn>=0.29.0          # ASGI server
twilio>=9.0.0            # SMS reminders
